requirements.txt
albumentations
numpy
PyYAML
rasterio
torch
torchvision
tqdm
